Simplify ScrollFinisher
It turned out that the previous way to calculate the target position of
the viewport for ListView.Contain case was half-broken anyway, which
becomes rather apparent now that the work on navigating to replied-to
events is ongoing. On the other hand, it looks like Qt has improved its
positioning algorithm even further so even if one simply calls
chatView.positionViewAtIndex() the positioning is already perfect or
close to perfect most of the time. So the logic is now simplified; newly
instantiated components don't have to ask scrollFinisher if they've been
called out, the finisher simply tries to call positionViewAtIndex()
several times at small intervals.